Change GeneticAlgorithm main loop logic
Old logic:
	Initialize creates first population and evaluates it
	Step spawns the next generation and scores it

New logic:
	Initialize creates first population
	Step scores the current population and spawns the next

With this logic, all of the GeneticAlgorithm members which return
scores or individuals operate on the current population. That is, the
population which we scored during the last step operation. When we call
Step again, it will score the pending population, swap it into the
current population, and spawn a new pending population.

Fixed BitVector::HammingDistance if the vector contains garbage bits at
the end.

Also added some documentation.